Switched to postMessage data protocol
We support only JSON encoded string requests in
raw HTTP requests, so that we're now able to distinguish
between null and "null" values for text fields.

The code still needs some mechanism to properly interpret
null values. We need a translation between the values
supplied in the HTTP request from the GUI to the values
passed to the OM instance.

Signed-off-by: Ionut Nicu <ionut.nicu@mindbit.ro>