Added arrayToOm() method in RestDataSource
This method translates the "data" array from the request
to the appopriate SQL database values and calls the setters
of the OM instance to populate it with data.

Currently we treat "" as NULL for text and numeric columns
that can be NULL. This should work for most of the cases
(even the inet postgresql type should be ok because propel
 considers its type to be varchar).

For the cases where this implementation is not good enough, the
method can be overridden by the user.

Signed-off-by: Ionut Nicu <ionut.nicu@mindbit.ro>