Improve doSave() logic in OmRequest
User-defined validators via validate() method and propel validators are
now checked independently. This allows for displaying all errors
simultaneously, regardless of the validators that generated them.

The real save() OM method call was moved to the __doSave() method that,
in turn, is called when the error list ($this->err) is empty. This way
additional tasks can be performed on the "real" save path (with all
validators successfully passed), for instance populating derived tables.

Signed-off-by: Radu Rendec <radu.rendec@mindbit.ro>