Add/Update consistency between view and controller
HTML::submitButton() generates a submit button *and* a hidden with a
value of "1". Since the "add" and "update" request types are not
forced from outside and typically HTML::submitButton() is used to
generate submit elements, it's safe to assume that the __add and
__update request inputs will always have a value of 1.

The SimpleFormRequest::decode() method has been changed to
additionally check for a "true" value of __add and __update, and thus
allowing the user to mask an Add or Update operation (by setting
the corresponding hidden field value to 0) when they want to force
another operation type (Remove, for instance).

Signed-off-by: Radu Rendec <radu.rendec@mindbit.ro>