Throw a more explicit remote exception
It looks like SmartClient doesn't look at the previous exception and
in the interface we see only "Exception on RMI server side" without
any detail of what was wrong on the remote end.

In order to get it in the interface we need to explicitly add the
previous exception message to the one we're throwing.

Signed-off-by: Ionut Nicu <ionut.nicu@mindbit.ro>