Fix AbstractErrorHandler::varDump() object recursion
Old code used to dynamically add the __varDump property to objects as a
recursion marker, so that later it could recognize the instance.

In php5 this can lead to unpredictable results with classes that
implement the __get and/or __set magic methods, because they actually
get called for __varDump propery reads/writes.

On the other hand, in php5 object variables are no longer the actual
instance, but a handle (pointer) to the actual instance. Therefore, we
can easily make a map of "seen" instances and then use the === operator
to identify them.

Signed-off-by: Radu Rendec <radu.rendec@mindbit.ro>