Template: strict parameter type for Block::setVariables()
Set a strict `array` parameter type for Block::setVariables(). The
function uses the `+` operator on arrays. Calling the function with
anything other than array generated a fatal error. By setting the
parameter type, an error is raised at the call site if the parameter
is not an array; the error can be handled by user-defined error
handlers.

Signed-off-by: Radu Rendec <radu.rendec@mindbit.ro>