Make SimpleFormRequest::actionFetch() throw if object does not exist
Most of the code assumes that OmRequest::$om is never null because it's
initialized by the constructor.

This change prevents SimpleFormRequest::actionFetch() from setting
$this->om to null when the requested object cannot be found in the
database. This avoids fatal errors if code that executes later tries to
invoke methods on the object that is held by $this-om.

An exception of ObjectNotFoundException type is explicitly thrown. If
the exception is caught, $this->om still holds a valid object and the
primary key is set in that object.

Signed-off-by: Radu Rendec <radu.rendec@mindbit.ro>